+++ gateway-js/src/schema-helper/addResolversToSchema.ts
